safari-534.53-branch/LayoutTests/fast/js/stack-overflow-arrity-catch-expected.txt

Test that if an arrity check causes a stack overflow, the exception goes to the right catch

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS gotRightCatch is true
PASS successfullyParsed is true

TEST COMPLETE
